refactor: CalendarService 내부 InstantRange 클래스를 calendar 패키지로 분리
